requirements.txt
Flask
Flask-Login
python-dotenv
exifread
reportlab
Pillow
folium
